requirements.txt
# Telegram Bot Core
aiogram==3.15.0
apscheduler==3.10.4

# Configuration
python-dotenv==1.0.1

# HTTP Client for API calls
httpx==0.27.0

# Caching
cachetools==5.5.0

# Date parsing
python-dateutil==2.9.0

# Testing
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
